New cleaner structure for project
