+++ MetaExchangeSowaLabs.Lib/CustomErrors/CantSellDesiredBtcException.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace MetaExchangeSowaLabs.Lib.CustomErrors
+{
+    public class CantSellDesiredBtcException : Exception
+    {
+        public  CantSellDesiredBtcException(decimal amount) 
+            : base($"Can't sell {amount} BTC!")
+        {
+        }
+        
+    }
+}
