+++ MetaExchangeSowaLabs.Lib/CustomErrors/EmptyOrderBookOrBalanceException.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace MetaExchangeSowaLabs.Lib.CustomErrors
+{
+    public class EmptyOrderBookOrBalanceException : Exception
+    {
+        public EmptyOrderBookOrBalanceException(string type)
+            : base($"No {type} available!")
+        {
+
+        }
+    }
+}
