+++ MetaExchangeSowaLabs.Lib/CustomErrors/UserHasNoMoneyOrBtcException.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace MetaExchangeSowaLabs.Lib.CustomErrors
+{
+    public class UserHasNoMoneyOrBtcException : Exception
+    {
+        public UserHasNoMoneyOrBtcException(string type)
+            : base($"User doesn't have enough {type}!")
+        {
+
+        }
+    }
+}
