+++ MetaExchangeSowaLabs.Lib/Entities/BidAskEntity.cs
@@ -1,0 +1,7 @@
+﻿namespace MetaExchangeSowaLabs.Lib.Entities
+{
+    public class BidAskEntity
+    {
+        public OrderEntity Order { get; set; }
+    }
+}
