+++ MetaExchangeSowaLabs.Lib/Entities/OrderBookBalanceEntity.cs
@@ -1,0 +1,9 @@
+﻿namespace MetaExchangeSowaLabs.Lib.Entities
+{
+    public class OrderBookBalanceEntity
+    {
+        public string OrderBookId { get; set; }
+        public decimal Bitcoin { get; set; }
+        public decimal Eur { get; set; }
+    }
+}
