+++ MetaExchangeSowaLabs.Lib/Enums/TypeOfOrderEnum.cs
@@ -1,0 +1,8 @@
+﻿namespace MetaExchangeSowaLabs.Lib.Enums
+{
+    public enum TypeOfOrderEnum
+    {
+        Buy = 1,
+        Sell = 2
+    }
+}
