+++ MetaExchangeSowaLabs.Lib/Helpers/FileIOHelper.cs
@@ -1,0 +1,14 @@
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+
+namespace MetaExchangeSowaLabs.Lib.Helpers
+{
+    public static class FileIOHelper
+    {
+        public static IEnumerable<string> ExtractNRows(string path, int numberOfRows)
+        {
+             return File.ReadLines(path).Take(numberOfRows);
+        }
+    }
+}
