+++ MetaExchangeSowaLabs.sln
@@ -4,6 +4,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MetaExchangeSowaLabs", "Met
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MetaExchangeSowaLabs.Tests", "MetaExchangeSowaLabs.Tests\MetaExchangeSowaLabs.Tests.csproj", "{00415C81-38D7-4B74-82ED-E0E6E8BE29BF}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MetaExchangeSowaLabs.Lib", "MetaExchangeSowaLabs.Lib\MetaExchangeSowaLabs.Lib.csproj", "{2041F5AB-B820-47B7-BB33-1791FB769FFB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -18,5 +20,9 @@ Global
 		{00415C81-38D7-4B74-82ED-E0E6E8BE29BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{00415C81-38D7-4B74-82ED-E0E6E8BE29BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{00415C81-38D7-4B74-82ED-E0E6E8BE29BF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2041F5AB-B820-47B7-BB33-1791FB769FFB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2041F5AB-B820-47B7-BB33-1791FB769FFB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2041F5AB-B820-47B7-BB33-1791FB769FFB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2041F5AB-B820-47B7-BB33-1791FB769FFB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
