--- MetaExchangeSowaLabs/CustomErrors/CantBuyDesiredBtcException.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace MetaExchangeSowaLabs.CustomErrors
-{
-    public class CantBuyDesiredBtcException : Exception
-    {
-        public  CantBuyDesiredBtcException(decimal amount) 
-            : base($"Can't buy {amount} BTC!")
-        {
-        }
-    }
-}
