--- MetaExchangeSowaLabs/CustomErrors/EmptyOrderBookOrBalanceException.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace MetaExchangeSowaLabs.CustomErrors
-{
-    public class EmptyOrderBookOrBalanceException : Exception
-    {
-        public EmptyOrderBookOrBalanceException(string type)
-            : base($"No {type} available!")
-        {
-
-        }
-    }
-}
