--- MetaExchangeSowaLabs/CustomErrors/UserHasNoMoneyOrBtcException.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace MetaExchangeSowaLabs.CustomErrors
-{
-    public class UserHasNoMoneyOrBtcException : Exception
-    {
-        public UserHasNoMoneyOrBtcException(string type)
-            : base($"User doesn't have enough {type}!")
-        {
-
-        }
-    }
-}
