--- MetaExchangeSowaLabs/Entities/BidAskEntity.cs
@@ -1,7 +1,0 @@
-﻿namespace MetaExchangeSowaLabs.Entities
-{
-    public class BidAskEntity
-    {
-        public OrderEntity Order { get; set; }
-    }
-}
