--- MetaExchangeSowaLabs/Entities/OrderBookBalanceEntity.cs
@@ -1,9 +1,0 @@
-﻿namespace MetaExchangeSowaLabs.Entities
-{
-    public class OrderBookBalanceEntity
-    {
-        public string OrderBookId { get; set; }
-        public decimal Bitcoin { get; set; }
-        public decimal Eur { get; set; }
-    }
-}
