--- MetaExchangeSowaLabs/Enums/TypeOfOrderEnum.cs
@@ -1,8 +1,0 @@
-﻿namespace MetaExchangeSowaLabs.Enums
-{
-    public enum TypeOfOrderEnum
-    {
-        Buy,
-        Sell
-    }
-}
