+++ MetaExchangeSowaLabs/MetaExchangeSowaLabs.csproj
@@ -18,4 +18,8 @@
       <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
     </ItemGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\MetaExchangeSowaLabs.Lib\MetaExchangeSowaLabs.Lib.csproj" />
+    </ItemGroup>
+
 </Project>
